Make notifier configurable
